requirements.txt
Babel==1.3
Django==1.8
M2Crypto==0.22.3
Passbook==1.0.0
Pillow==2.7.0
PyJWT==1.0.0
PyYAML==3.11
amqp==1.4.6
anyjson==0.3.3
argparse==1.2.1
backport-collections==0.1
beautifulsoup4==4.3.2
billiard==3.3.0.19
celery==3.1.17
cffi==0.9.2
contextlib2==0.4.0
coverage==3.7.1
cryptography==0.8.1
django-appconf==1.0.1
django-bootstrap-form==3.1
django-celery==3.1.16
django-ckeditor-updated==4.4.4
django-compressor==1.4
django-countries==3.1.1
django-datetime-widget==0.9.3
django-dbbackup==2.0.4
django-easy-maps==0.9
django-extensions==1.5.0
django-form-utils==1.0.2
django-grappelli==2.6.3
django-htmlmin==0.7.0
django-jenkins==0.17.0
django-lineage==0.2.0
django-model-utils==2.2
django-post-office==1.1.1
django-statici18n==1.1.1
django-timedeltafield==0.7.3
djangorestframework==3.1.0
drf-extensions==0.2.7
ecdsa==0.13
enum34==1.0.4
factory-boy==2.4.0
geopy==1.8.1
gunicorn==19.2.0
html5lib==0.999
jsonfield==1.0.3
kombu==3.0.24
mock==1.0.1
newrelic==2.42.0.35
nose==1.3.4
paramiko==1.15.2
psycopg2==2.5.4
pyasn1==0.1.7
pycparser==2.10
pycrypto==2.6.1
pyen==2.3
pysftp==0.2.8
python-dateutil==2.4.0
python-memcached==1.53
pytz==2015.2
rauth==0.7.1
raven==5.1.1
requests==2.5.1
six==1.9.0
sorl-thumbnail==12.2
suds==0.4
vcrpy==1.4.1
wrapt==1.10.4
wsgiref==0.1.2
flower==0.8.1
facebook-sdk==0.4.0
